requirements.txt
firebase_admin==6.0.1
joblib==1.2.0
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.1
plotly==5.11.0
python-dotenv==0.21.0
scikit_learn==1.1.3
streamlit==1.14.1
streamlit_autorefresh==0.0.1
tensorflow==2.10.0
wordcloud==1.8.2.2
